# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggeregator",
        "description": "Laravel News Aggeragator API",
        "contact": {
            "email": "amr@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/preferences": {
            "get": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Get user preferences",
                "description": "Get list of user preferences",
                "operationId": "f1eb50f3bfb30dde3ed40dac8f0e1780",
                "responses": {
                    "200": {
                        "description": "User preferences retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Preference"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Set user preference",
                "description": "Create or update user preference",
                "operationId": "1a3e7b639a919481e5e3f12364358007",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "preferencable_id",
                                    "preferencable_type"
                                ],
                                "properties": {
                                    "preferencable_id": {
                                        "description": "ID of the preferencable",
                                        "type": "integer"
                                    },
                                    "preferencable_type": {
                                        "description": "Type of the preferencable entity",
                                        "type": "string",
                                        "enum": [
                                            "category",
                                            "author",
                                            "source"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Preference created/updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Preference"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Author": {
                "title": "Author Resource",
                "description": "Author resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category Resource",
                "description": "Category resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Programming"
                    }
                },
                "type": "object"
            },
            "Preference": {
                "title": "Preference Resource",
                "description": "Preference resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "preferencable_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "preferencable_type": {
                        "type": "string",
                        "example": "App\\\\Models\\\\Category"
                    },
                    "preferencable": {
                        "description": "Preferencable object",
                        "type": "object",
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Category"
                            },
                            {
                                "$ref": "#/components/schemas/Author"
                            },
                            {
                                "$ref": "#/components/schemas/Source"
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "Source": {
                "title": "Source Resource",
                "description": "Source resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Local"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}